Refactored files for organization
